requirements.txt
aiogram==3.4.1
python-dotenv==1.0.0
apscheduler==3.10.4
aiosqlite==0.19.0
asyncpg==0.29.0
aiohttp==3.9.5
